stuff again

+++ src/com/arinerron/forux/core/ResourceManager.java
@@ -22,6 +22,8 @@ public class ResourceManager {
     private File images = null;
     private File audio = null;
 
+    private boolean working = false;
+
     protected ResourceManager(Game game) {
         this.game = game;
         this.home = new File(new File(System.getProperty("user.home")), "." + this.getGame().getFileName());
@@ -45,7 +47,7 @@ public class ResourceManager {
             if(!home.exists())
                 this.getGame().setFirstRun(true);
             home.mkdirs();
-            if(!config.exists())
+            if(!config.exists() && !working)
                 writeFile(config, "{}");
             screenshots.mkdirs();
             logs.mkdirs();
@@ -59,16 +61,7 @@ public class ResourceManager {
     }
 
     public String readFile(File file) throws IOException {
-        if(!home.exists())
-            this.getGame().setFirstRun(true);
-        home.mkdirs();
-        screenshots.mkdirs();
-        logs.mkdirs();
-        log.createNewFile();
-        resources.mkdirs();
-        images.mkdirs();
-        audio.mkdirs();
-
+        check();
         if(!file.exists())
             file.createNewFile();
 
@@ -86,7 +79,9 @@ public class ResourceManager {
     }
 
     public void writeFile(File file, Object contents) throws IOException {
+        working = true;
         check();
+        working = false;
 
         if(!file.exists())
             file.createNewFile();
